Creation of separate losses and callbacks modules
