+++ klax/__init__.py
@@ -1,6 +1,10 @@
-from . import nn as nn, typing as typing, wrappers as wrappers
+from . import (
+    callbacks as callbacks,
+    losses as losses,
+    nn as nn,
+    wrappers as wrappers
+)
 from ._training import (
-    CallbackArgs as CallbackArgs,
     dataloader as dataloader,
     fit as fit
 )
